Fix StepCount key from Int to Long to support epoch time

+++ app/src/main/java/com/hci_g1/amelior/databases/dao/StepCountDao.kt
@@ -14,13 +14,13 @@ interface StepCountDao
 
     @Transaction
     @Query("SELECT EXISTS(SELECT * FROM StepCount WHERE key = :key LIMIT 1)")
-    fun step_count_exists(key: Int): Boolean
+    fun step_count_exists(key: Long): Boolean
 
     @Transaction
     @Query("SELECT * FROM StepCount WHERE key = :key LIMIT 1")
-    suspend fun get_step_count(key: Int): StepCount
+    suspend fun get_step_count(key: Long): StepCount
 
     @Transaction
     @Query("SELECT * FROM StepCount WHERE key = :key LIMIT 1")
-    fun get_step_count_now(key: Int): StepCount
+    fun get_step_count_now(key: Long): StepCount
 }
